Changed VAT number to be long. Added script to start downloading of ehf companies compliance list.

+++ messages/src/main/java/com/thundashop/core/getshop/data/EhfComplientCompany.java
@@ -12,6 +12,6 @@ import com.thundashop.core.common.DataCommon;
  * @author ktonder
  */
 public class EhfComplientCompany extends DataCommon {
-    public int vatNumber;
+    public long vatNumber;
     public String name;
 }
